feat(theme): 添加新建主题功能并重构主进程代码
- 在 zh-CN.json 中添加新主题相关翻译
- 重构 main.js，将窗口创建和文件操作移至 utils 模块
- 添加 adm-zip 依赖用于文件压缩
- 更新 CSS 样式，包括全局按钮样式和格式化
- 在 new-file.js 中添加新主题按钮事件处理
- 注释掉主菜单中的编辑按钮相关代码
