+++ package.json
@@ -10,6 +10,7 @@
   "author": "Frank Steven",
   "license": "MIT",
   "dependencies": {
+    "adm-zip": "^0.5.16",
     "electron": "^25.9.8"
   }
 }
