+++ src/templates/css/global.css
@@ -1,10 +1,28 @@
 * {
   user-select: none;
 }
+
 body {
   margin: 0;
   background-color: var(--bg0);
   color: var(--fg0);
   height: 100vh;
   width: 100vw;
+}
+
+.big-flex-btn {
+  margin: 10px;
+  padding: 10px;
+  border: none;
+  border-radius: 5px;
+  background-color: var(--bg1);
+  color: var(--fg1);
+  font-size: 1.2em;
+  cursor: pointer;
+  width: 150px;
+  height: 100px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
 }
