+++ src/templates/css/main-menu.css
@@ -2,6 +2,7 @@
   height: 100vh;
   display: flex;
 }
+
 #main-menu-sidebar {
   background-color: var(--bg1);
   height: 100%;
@@ -9,6 +10,7 @@
   display: flex;
   flex-direction: column;
 }
+
 #main-menu-sidebar button {
   color: var(--fg1);
   background-color: var(--bg1);
@@ -16,46 +18,54 @@
   border: none;
   font-size: var(--fs4);
 }
+
 #main-menu-sidebar button:hover {
   background-color: var(--bg2);
 }
+
 #main-menu-content {
   width: 100%;
   overflow: auto;
 }
+
 #main-menu-content div {
   margin: 30px;
 }
+
 #main-menu-content-start-buttons {
-    display: flex;
-    gap: 20px;
-    align-items: center;
-    justify-content: center;
+  display: flex;
+  gap: 20px;
+  align-items: center;
+  justify-content: center;
 }
+
 #main-menu-content-start-buttons button {
-    background-color: var(--bg1);
-    border: none;
-    color: var(--fg1);
-    width: 150px;
-    height: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    gap: 10px;
-    font-size: var(--fs4);
-    padding: 10px 20px;
-    border-radius: 5px;
-    cursor: pointer;
+  background-color: var(--bg1);
+  border: none;
+  color: var(--fg1);
+  width: 150px;
+  height: 100px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  gap: 10px;
+  font-size: var(--fs4);
+  padding: 10px 20px;
+  border-radius: 5px;
+  cursor: pointer;
 }
+
 #main-menu-content-start-buttons button:hover {
-    background-color: var(--bg2);
+  background-color: var(--bg2);
 }
-#main-menu-content-start-buttons button span{
-    font-size: var(--fs5);
+
+#main-menu-content-start-buttons button span {
+  font-size: var(--fs5);
 }
+
 #main-menu-content-start-buttons button img {
-    display: block;
-    width: 40px;
-    height: 40px;
+  display: block;
+  width: 40px;
+  height: 40px;
 }
