+++ src/templates/css/new-file.css
@@ -1,60 +1,61 @@
 #new-file {
-    background-color: var(--bg0);
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    width: 100%;
+  background-color: var(--bg0);
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
 }
 
 #new-file>div {
-    margin: 0 40px;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
+  margin: 0 40px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 #new-file-theme-select {
-    width: 100%;
-    height: calc(100% - 200px);
-    display: flex;
-    overflow-y: auto;
-    flex-wrap: wrap;
-    align-items: flex-start;
-    justify-content: flex-start;
+  width: 100%;
+  height: calc(100% - 200px);
+  display: flex;
+  overflow-y: auto;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  justify-content: flex-start;
 }
 
 ::-webkit-scrollbar {
-    display: none;
+  display: none;
 }
 
-#new-file-theme-select > div {
-    display: flex;
-    flex-wrap: wrap;
-    height: 100%;
-    justify-content: left;
-    align-items: center;
-    flex-direction: column;
+#new-file-theme-select>div {
+  display: flex;
+  flex-wrap: wrap;
+  height: 100%;
+  justify-content: left;
+  align-items: center;
+  flex-direction: column;
 }
 
+/* NOTE: 请请用 class="big-flex-btn" in global.css 而不是这个 */
 #new-file-theme-select button {
-    margin: 10px;
-    padding: 10px;
-    border: none;
-    border-radius: 5px;
-    background-color: var(--bg1);
-    color: var(--fg1);
-    font-size: 1.2em;
-    cursor: pointer;
-    width: 150px;
-    height: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
+  margin: 10px;
+  padding: 10px;
+  border: none;
+  border-radius: 5px;
+  background-color: var(--bg1);
+  color: var(--fg1);
+  font-size: 1.2em;
+  cursor: pointer;
+  width: 150px;
+  height: 100px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
 }
 
-#new-file-theme-select > div > button > img {
-    display: block;
-    width: 40px;
-    height: 40px;
+#new-file-theme-select>div>button>img {
+  display: block;
+  width: 40px;
+  height: 40px;
 }
