+++ src/templates/main-menu.html
@@ -27,13 +27,13 @@
               />
               <span id="main-menu-content-start-buttons-new-text"></span>
             </button>
-            <button id="main-menu-content-start-buttons-new-edit">
+            <!-- <button id="main-menu-content-start-buttons-new-edit">
               <img
                 id="main-menu-content-start-buttons-new-edit-icon"
                 src="../asset/imgs/zhan.png"
               />
               <span id="main-menu-content-start-buttons-new-edit-text"></span>
-            </button>
+            </button> -->
             <button id="main-menu-content-start-buttons-open">
               <img
                 id="main-menu-content-start-buttons-open-icon"
