+++ src/templates/new-file.html
@@ -14,9 +14,9 @@
         <p id="new-file-description"></p>
         <div id="new-file-theme-select">
           <div>
-            <button id="new-file-theme-select-new-theme">
+            <button id="new-file-theme-select-new-theme" class="big-flex-btn">
               <img id="new-file-theme-select-new-theme-icon" src="../asset/imgs/zhan.png"></img>
-              <span id="new-file-theme-select-new-theme-text">new</span>
+              <span id="new-file-theme-select-new-theme-text"></span>
             </button>
             <button>
               <span>2nd</span>
