+++ src/templates/new-theme.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title id="new-theme-title"></title>
+    <link id="theme-stylesheet" rel="stylesheet" href="" />
+    <link rel="stylesheet" href="./css/global.css" />
+    <link rel="stylesheet" href="./css/new-theme.css" />
+  </head>
+
+  <body>
+    <div id="new-theme">
+      <div>
+        <h1 id="new-theme-header"></h1>
+        <p id="new-theme-description"></p>
+        <div id="new-theme-foreground">
+          <p id="new-theme-foreground-description"></p>
+          <div>
+            <button id="new-theme-foreground-import" class="big-flex-btn">
+              <img id="new-theme-foreground-import-icon" src="../asset/imgs/zhan.png"></img>
+              <span id="new-theme-foreground-import-text"></span>
+            </button>
+            <button class="big-flex-btn">
+              <span>2nd</span>
+            </button>
+            <button class="big-flex-btn"></button>
+            <button class="big-flex-btn"></button>
+            <button class="big-flex-btn"></button>
+            <button class="big-flex-btn"></button>
+            <button class="big-flex-btn"></button>
+          </div>
+        </div>
+        <div id="new-theme-background">
+          <p id="new-theme-background-description"></p>
+        </div>
+      </div>
+    </div>
+  </body>
+
+  <script src="./js/global.js"></script>
+  <script src="./js/new-theme.js"></script>
+</html>
